Respond to code review
- Update Guide title to "Visual identity"
- Delete -old files
- Move filter functions, including color contrast, to config/
- Fix line breaks, styles, and URLs across pages
- Add note to README about how to link files for download
