--- assets/brand/styles/_color-matrix-old.scss
@@ -1,170 +1,0 @@
-// "Marge" is a cross between USWDS's "Medium" and "Large" breakpoints.
-// It's the equivalent of Bootstrap's "Medium" size.
-$marge-screen: 992px;
-
-// The WAI-ARIA specification states the following about aria-hidden:
-//
-//   Authors MAY, with caution, use aria-hidden to hide visibly rendered
-//   content from assistive technologies only if the act of hiding this
-//   content is intended to improve the experience for users of assistive
-//   technologies by removing redundant or extraneous content. Authors
-//   using aria-hidden to hide visible content from screen readers MUST
-//   ensure that identical or equivalent meaning and functionality is
-//   exposed to assistive technologies.
-//
-// Unfortunately, however, the U.S. Web Design Standards forcibly set
-// anything with aria-hidden to display: none, making it difficult to
-// address the above use case. The following rule allows us to override
-// this default and create content that is visible to sighted users but
-// irrelevant to non-sighted users.
-.usa-sr-invisible[aria-hidden="true"] {
-  display: block !important;
-}
-
-.usa-matrix-foreground-white {
-  color: $color-white;
-
-  // https://css-tricks.com/adding-stroke-to-web-text/
-  text-shadow:
-   -1px -1px 0 #000,
-    1px -1px 0 #000,
-    -1px 1px 0 #000,
-     1px 1px 0 #000;
-}
-
-.usa-matrix-foreground-light {
-  color: $color-light;
-}
-
-.usa-matrix-foreground-bright {
-  color: $color-bright;
-}
-
-.usa-matrix-foreground-medium {
-  color: $color-medium;
-}
-
-.usa-matrix-foreground-dark {
-  color: $color-dark;
-}
-
-.usa-matrix-foreground-black {
-  color: $color-black;
-}
-
-.usa-matrix td {
-  vertical-align: top;
-}
-
-.usa-matrix-square {
-  width: 4em;
-  height: 4em;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  strong {
-    text-shadow: none;
-  }
-}
-
-.usa-matrix-symbol-definitions {
-  display: none;
-}
-
-.usa-matrix-legend {
-  position: relative;
-
-  svg {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 35px;
-    height: 35px;
-  }
-
-  p {
-    margin-left: 45px;
-    font-size: 0.75em;
-  }
-}
-
-td[scope="row"] > div {
-  display: flex;
-
-  .usa-matrix-desc {
-    flex: 1;
-  }
-}
-
-@media (min-width: $marge-screen) and (max-width: "#{$large-screen - 1}") {
-  table th, table td {
-    padding: 0.5em;
-  }
-
-  .usa-matrix-desc {
-    font-size: 0.75em;
-  }
-
-  td[scope="row"] > div {
-    flex-direction: column;
-
-    .usa-matrix-desc {
-      padding-top: 0.5em;
-    }
-  }
-}
-
-@media (min-width: $large-screen) {
-  td[scope="row"] > div {
-    .usa-matrix-desc {
-      padding-left: 0.5em;
-    }
-  }
-}
-
-@media (min-width: $marge-screen) {
-  .usa-matrix-color-combo-description {
-    @include sr-only();
-  }
-}
-
-@media (max-width: "#{$marge-screen - 1}") {
-  .usa-matrix-legend {
-    display: none;
-  }
-
-  table.usa-matrix {
-    display: block;
-
-    thead {
-      display: none;
-    }
-
-    tbody {
-      display: block;
-
-      tr {
-        display: block;
-
-        td.usa-matrix-valid-color-combo {
-          display: flex;
-          align-items: center;
-
-          .usa-matrix-color-combo-description {
-            flex: 1;
-            padding-left: 1em;
-          }
-        }
-
-        td.usa-matrix-invalid-color-combo {
-          display: none;
-        }
-
-        td[scope="row"] {
-          display: none;
-        }
-      }
-    }
-  }
-}
